fix: promote not moving things in filter mode
